requirements.txt
futu-api==10.0.6008
PyYAML==6.0.2
